requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
altair==5.1.2
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
appnope==0.1.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.7.2
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
Babel==2.13.1
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.2
bleach==6.1.0
blinker==1.7.0
blis==0.7.11
bokeh==3.3.4
build==0.9.0
cachetools==5.3.2
catalogue==2.0.10
certifi==2022.12.7
cffi==1.15.1
chardet==5.2.0
charset-normalizer==2.0.12
chroma-hnswlib==0.7.3
chromadb==0.4.20
click==8.1.7
cloudpathlib==0.16.0
coloredlogs==15.0.1
comm==0.1.4
confection==0.1.4
contourpy==1.2.0
cryptography==3.4.6
curated-tokenizers==0.0.8
curated-transformers==0.1.1
cycler==0.12.1
cymem==2.0.8
Cython==0.29.36
dataclasses-json==0.6.1
datasets==2.16.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.7
distro==1.8.0
effdet==0.4.1
emoji==2.8.0
en-core-web-trf @ https://github.com/explosion/spacy-models/releases/download/en_core_web_trf-3.7.3/en_core_web_trf-3.7.3-py3-none-any.whl#sha256=f72abb34bdf174876bd4267b29b2501677e605e0a251fdc56c163003182ed68b
et-xmlfile==1.1.0
exceptiongroup==1.1.3
executing==2.0.1
faiss-cpu==1.7.4
fastapi==0.105.0
fastjsonschema==2.18.1
filelock==3.13.1
filetype==1.2.0
Flask==3.0.0
Flask-Cors==4.0.0
flatbuffers==23.5.26
fonttools==4.44.3
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.10.0
gevent==23.9.1
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.25.2
googleapis-common-protos==1.62.0
greenlet==3.0.1
grpcio==1.60.0
h11==0.14.0
hdbscan==0.8.33
html5lib==1.1
httpcore==1.0.2
httptools==0.6.1
httpx==0.25.1
huggingface-hub==0.19.4
humanfriendly==10.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.1
iopath==0.1.10
ipykernel==6.26.0
ipympl==0.9.3
ipython==8.17.2
ipython-genutils==0.2.0
ipywidgets==8.1.1
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonlines==4.0.0
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter-events==0.8.0
jupyter-lsp==2.2.0
jupyter_client==8.5.0
jupyter_core==5.5.0
jupyter_server==2.9.1
jupyter_server_terminals==0.4.4
jupyterlab==4.0.7
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.25.0
kiwisolver==1.4.5
kneed==0.8.5
kubernetes==28.1.0
langchain==0.0.341
langchain-benchmarks==0.0.10
langchain-core==0.0.6
langchain-experimental==0.0.37
langchainhub==0.1.13
langcodes==3.3.0
langdetect==1.0.9
langsmith==0.0.86
layoutparser==0.3.4
linkify-it-py==2.0.3
llvmlite==0.41.1
lxml==4.9.3
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.1
matplotlib-inline==0.1.6
mdit-py-plugins==0.4.0
mdurl==0.1.2
mistune==3.0.2
mlxtend==0.23.1
mmh3==4.0.1
monotonic==1.6
mpmath==1.3.0
msg-parser==1.2.0
multidict==6.0.4
multipart==0.2.4
multiprocess==0.70.15
murmurhash==1.0.10
mypy-extensions==1.0.0
nbclient==0.8.0
nbconvert==7.10.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.2.1
nltk==3.8.1
notebook_shim==0.2.3
numba==0.58.1
numpy==1.26.1
oauthlib==3.2.2
olefile==0.46
omegaconf==2.3.0
onnx==1.15.0
onnxruntime==1.15.1
openai==1.3.5
opencv-python==4.8.1.78
openpyxl==3.1.2
opentelemetry-api==1.21.0
opentelemetry-exporter-otlp-proto-common==1.21.0
opentelemetry-exporter-otlp-proto-grpc==1.21.0
opentelemetry-instrumentation==0.42b0
opentelemetry-instrumentation-asgi==0.42b0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-proto==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-semantic-conventions==0.42b0
opentelemetry-util-http==0.42b0
overrides==7.4.0
packaging==21.3
pandas==2.1.2
pandocfilters==1.5.0
panel==1.3.8
param==2.0.2
parso==0.8.3
pdf2image==1.16.3
pdfminer.six==20221105
pdfplumber==0.10.3
pdfservices-sdk==2.3.0
pep517==0.13.0
pexpect==4.8.0
Pillow==10.1.0
platformdirs==3.11.0
plotly==5.18.0
polling==0.3.2
polling2==0.5.0
portalocker==2.8.2
posthog==3.1.0
preshed==3.0.9
prometheus-client==0.18.0
prompt-toolkit==3.0.39
protobuf==4.25.0
psutil==5.9.6
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pulsar-client==3.3.0
pure-eval==0.2.2
pyarrow==14.0.1
pyarrow-hotfix==0.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycocotools==2.0.7
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
pydeck==0.8.1b0
Pygments==2.14.0
PyJWT==2.4.0
PyMuPDF==1.23.6
PyMuPDFb==1.23.6
pynndescent==0.5.11
pypandoc==1.12
pyparsing==3.0.9
pypdf==3.17.1
PyPDF2==2.12.1
pypdfium2==4.24.0
PyPika==0.48.9
pysbd==0.3.4
pytesseract==0.3.10
python-dateutil==2.8.2
python-docx==1.1.0
python-dotenv==1.0.0
python-iso639==2023.6.15
python-json-logger==2.0.7
python-magic==0.4.27
python-mnist==0.7
python-multipart==0.0.6
python-pptx==0.6.23
pytz==2023.3.post1
pyviz_comms==3.0.1
PyYAML==6.0
pyzmq==25.1.1
ragas==0.0.22
rapidfuzz==3.5.2
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.6.0
rpds-py==0.10.6
rsa==4.9
safetensors==0.4.0
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.13.1
Send2Trash==1.8.2
sentence-transformers==2.3.1
sentencepiece==0.1.99
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
spacy==3.7.2
spacy-curated-transformers==0.2.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.22
srsly==2.4.8
stack-data==0.6.3
starlette==0.27.0
streamlit==1.28.2
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
terminado==0.17.1
thinc==8.2.1
thinc-apple-ops==0.1.4
threadpoolctl==3.2.0
tiktoken==0.5.1
timm==0.9.11
tinycss2==1.2.1
tokenizers==0.15.0
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==2.1.1
torchvision==0.16.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.13.0
transformers==4.35.2
typer==0.9.0
types-python-dateutil==2.8.19.14
types-requests==2.31.0.10
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
uc-micro-py==1.0.3
umap-learn==0.5.5
unstructured==0.11.0
unstructured-inference==0.7.15
unstructured.pytesseract==0.3.12
uri-template==1.3.0
urllib3==1.26.13
uvicorn==0.24.0.post1
uvloop==0.19.0
validators==0.22.0
wasabi==1.1.2
watchfiles==0.21.0
wcwidth==0.2.9
weasel==0.3.4
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
websockets==12.0
Werkzeug==3.0.1
widgetsnbextension==4.0.9
wrapt==1.16.0
xlrd==2.0.1
XlsxWriter==3.1.9
xxhash==3.4.1
xyzservices==2023.10.1
yarl==1.9.2
zipp==3.17.0
zope.event==5.0
zope.interface==6.1
